refactor: move RealmRepository and RealmPolicy into ferriskey-domain
