+++ libs/ferriskey-domain/src/realm/mod.rs
@@ -1,3 +1,5 @@
+pub mod ports;
+
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use std::str::FromStr;
